Struct aws_sdk_codecommit::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for CodeCommit_20150413.

This client allows ergonomic access to a CodeCommit_20150413-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn associate_approval_rule_template_with_repository(
&self
) -> AssociateApprovalRuleTemplateWithRepository<C, M, R>

pub fn batch_associate_approval_rule_template_with_repositories(
&self
) -> BatchAssociateApprovalRuleTemplateWithRepositories<C, M, R>

pub fn batch_describe_merge_conflicts(
&self
) -> BatchDescribeMergeConflicts<C, M, R>

pub fn batch_disassociate_approval_rule_template_from_repositories(
&self
) -> BatchDisassociateApprovalRuleTemplateFromRepositories<C, M, R>

pub fn batch_get_commits(&self) -> BatchGetCommits<C, M, R>

pub fn batch_get_repositories(&self) -> BatchGetRepositories<C, M, R>

pub fn create_approval_rule_template(
&self
) -> CreateApprovalRuleTemplate<C, M, R>

pub fn create_branch(&self) -> CreateBranch<C, M, R>

pub fn create_commit(&self) -> CreateCommit<C, M, R>

pub fn create_pull_request(&self) -> CreatePullRequest<C, M, R>

pub fn create_pull_request_approval_rule(
&self
) -> CreatePullRequestApprovalRule<C, M, R>

pub fn create_repository(&self) -> CreateRepository<C, M, R>

pub fn create_unreferenced_merge_commit(
&self
) -> CreateUnreferencedMergeCommit<C, M, R>

pub fn delete_approval_rule_template(
&self
) -> DeleteApprovalRuleTemplate<C, M, R>

pub fn delete_branch(&self) -> DeleteBranch<C, M, R>

pub fn delete_comment_content(&self) -> DeleteCommentContent<C, M, R>

pub fn delete_file(&self) -> DeleteFile<C, M, R>

pub fn delete_pull_request_approval_rule(
&self
) -> DeletePullRequestApprovalRule<C, M, R>

pub fn delete_repository(&self) -> DeleteRepository<C, M, R>

pub fn describe_merge_conflicts(&self) -> DescribeMergeConflicts<C, M, R>

pub fn describe_pull_request_events(&self) -> DescribePullRequestEvents<C, M, R>

pub fn disassociate_approval_rule_template_from_repository(
&self
) -> DisassociateApprovalRuleTemplateFromRepository<C, M, R>

pub fn evaluate_pull_request_approval_rules(
&self
) -> EvaluatePullRequestApprovalRules<C, M, R>

pub fn get_approval_rule_template(&self) -> GetApprovalRuleTemplate<C, M, R>

pub fn get_blob(&self) -> GetBlob<C, M, R>

pub fn get_branch(&self) -> GetBranch<C, M, R>

pub fn get_comment(&self) -> GetComment<C, M, R>

pub fn get_comment_reactions(&self) -> GetCommentReactions<C, M, R>

pub fn get_comments_for_compared_commit(
&self
) -> GetCommentsForComparedCommit<C, M, R>

pub fn get_comments_for_pull_request(
&self
) -> GetCommentsForPullRequest<C, M, R>

pub fn get_commit(&self) -> GetCommit<C, M, R>

pub fn get_differences(&self) -> GetDifferences<C, M, R>

pub fn get_file(&self) -> GetFile<C, M, R>

pub fn get_folder(&self) -> GetFolder<C, M, R>

pub fn get_merge_commit(&self) -> GetMergeCommit<C, M, R>

pub fn get_merge_conflicts(&self) -> GetMergeConflicts<C, M, R>

pub fn get_merge_options(&self) -> GetMergeOptions<C, M, R>

pub fn get_pull_request(&self) -> GetPullRequest<C, M, R>

pub fn get_pull_request_approval_states(
&self
) -> GetPullRequestApprovalStates<C, M, R>

pub fn get_pull_request_override_state(
&self
) -> GetPullRequestOverrideState<C, M, R>

pub fn get_repository(&self) -> GetRepository<C, M, R>

pub fn get_repository_triggers(&self) -> GetRepositoryTriggers<C, M, R>

pub fn list_approval_rule_templates(&self) -> ListApprovalRuleTemplates<C, M, R>

pub fn list_associated_approval_rule_templates_for_repository(
&self
) -> ListAssociatedApprovalRuleTemplatesForRepository<C, M, R>

pub fn list_branches(&self) -> ListBranches<C, M, R>

pub fn list_pull_requests(&self) -> ListPullRequests<C, M, R>

pub fn list_repositories(&self) -> ListRepositories<C, M, R>

pub fn list_repositories_for_approval_rule_template(
&self
) -> ListRepositoriesForApprovalRuleTemplate<C, M, R>

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

pub fn merge_branches_by_fast_forward(
&self
) -> MergeBranchesByFastForward<C, M, R>

pub fn merge_branches_by_squash(&self) -> MergeBranchesBySquash<C, M, R>

pub fn merge_branches_by_three_way(&self) -> MergeBranchesByThreeWay<C, M, R>

pub fn merge_pull_request_by_fast_forward(
&self
) -> MergePullRequestByFastForward<C, M, R>

pub fn merge_pull_request_by_squash(&self) -> MergePullRequestBySquash<C, M, R>

pub fn merge_pull_request_by_three_way(
&self
) -> MergePullRequestByThreeWay<C, M, R>

pub fn override_pull_request_approval_rules(
&self
) -> OverridePullRequestApprovalRules<C, M, R>

pub fn post_comment_for_compared_commit(
&self
) -> PostCommentForComparedCommit<C, M, R>

pub fn post_comment_for_pull_request(
&self
) -> PostCommentForPullRequest<C, M, R>

pub fn post_comment_reply(&self) -> PostCommentReply<C, M, R>

pub fn put_comment_reaction(&self) -> PutCommentReaction<C, M, R>

pub fn put_file(&self) -> PutFile<C, M, R>

pub fn put_repository_triggers(&self) -> PutRepositoryTriggers<C, M, R>

pub fn tag_resource(&self) -> TagResource<C, M, R>

pub fn test_repository_triggers(&self) -> TestRepositoryTriggers<C, M, R>

pub fn untag_resource(&self) -> UntagResource<C, M, R>

pub fn update_approval_rule_template_content(
&self
) -> UpdateApprovalRuleTemplateContent<C, M, R>

pub fn update_approval_rule_template_description(
&self
) -> UpdateApprovalRuleTemplateDescription<C, M, R>

pub fn update_approval_rule_template_name(
&self
) -> UpdateApprovalRuleTemplateName<C, M, R>

pub fn update_comment(&self) -> UpdateComment<C, M, R>

pub fn update_default_branch(&self) -> UpdateDefaultBranch<C, M, R>

pub fn update_pull_request_approval_rule_content(
&self
) -> UpdatePullRequestApprovalRuleContent<C, M, R>

pub fn update_pull_request_approval_state(
&self
) -> UpdatePullRequestApprovalState<C, M, R>

pub fn update_pull_request_description(
&self
) -> UpdatePullRequestDescription<C, M, R>

pub fn update_pull_request_status(&self) -> UpdatePullRequestStatus<C, M, R>

pub fn update_pull_request_title(&self) -> UpdatePullRequestTitle<C, M, R>

pub fn update_repository_description(
&self
) -> UpdateRepositoryDescription<C, M, R>

pub fn update_repository_name(&self) -> UpdateRepositoryName<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self