Enum aws_sdk_codecommit::SdkError

pub enum SdkError<E, R = Response> {
    ConstructionFailure(Box<dyn Error + Send + Sync + 'static, Global>),
    DispatchFailure(Box<dyn Error + Send + Sync + 'static, Global>),
    ResponseError {
        err: Box<dyn Error + Send + Sync + 'static, Global>,
        raw: R,
    },
    ServiceError {
        err: E,
        raw: R,
    },
}

Failed SDK Result

Variants

ConstructionFailure(Box<dyn Error + Send + Sync + 'static, Global>)

The request failed during construction. It was not dispatched over the network.

Tuple Fields of ConstructionFailure

0: Box<dyn Error + Send + Sync + 'static, Global>

DispatchFailure(Box<dyn Error + Send + Sync + 'static, Global>)

The request failed during dispatch. An HTTP response was not received. The request MAY have been sent.

Tuple Fields of DispatchFailure

0: Box<dyn Error + Send + Sync + 'static, Global>

ResponseError

A response was received but it was not parseable according the the protocol (for example the server hung up while the body was being read)

Fields of ResponseError

err: Box<dyn Error + Send + Sync + 'static, Global>raw: R

ServiceError

An error response was received from the service

Fields of ServiceError

err: Eraw: R