Module aws_sdk_codecommit::error

Modules

actor_does_not_exist_exception

See ActorDoesNotExistException

approval_rule_content_required_exception

See ApprovalRuleContentRequiredException

approval_rule_does_not_exist_exception

See ApprovalRuleDoesNotExistException

approval_rule_name_already_exists_exception

See ApprovalRuleNameAlreadyExistsException

approval_rule_name_required_exception

See ApprovalRuleNameRequiredException

approval_rule_template_content_required_exception

See ApprovalRuleTemplateContentRequiredException

approval_rule_template_does_not_exist_exception

See ApprovalRuleTemplateDoesNotExistException

approval_rule_template_in_use_exception

See ApprovalRuleTemplateInUseException

approval_rule_template_name_already_exists_exception

See ApprovalRuleTemplateNameAlreadyExistsException

approval_rule_template_name_required_exception

See ApprovalRuleTemplateNameRequiredException

approval_state_required_exception

See ApprovalStateRequiredException

author_does_not_exist_exception

See AuthorDoesNotExistException

before_commit_id_and_after_commit_id_are_same_exception

See BeforeCommitIdAndAfterCommitIdAreSameException

blob_id_does_not_exist_exception

See BlobIdDoesNotExistException

blob_id_required_exception

See BlobIdRequiredException

branch_does_not_exist_exception

See BranchDoesNotExistException

branch_name_exists_exception

See BranchNameExistsException

branch_name_is_tag_name_exception

See BranchNameIsTagNameException

branch_name_required_exception

See BranchNameRequiredException

cannot_delete_approval_rule_from_template_exception

See CannotDeleteApprovalRuleFromTemplateException

cannot_modify_approval_rule_from_template_exception

See CannotModifyApprovalRuleFromTemplateException

client_request_token_required_exception

See ClientRequestTokenRequiredException

comment_content_required_exception

See CommentContentRequiredException

comment_content_size_limit_exceeded_exception

See CommentContentSizeLimitExceededException

comment_deleted_exception

See CommentDeletedException

comment_does_not_exist_exception

See CommentDoesNotExistException

comment_id_required_exception

See CommentIdRequiredException

comment_not_created_by_caller_exception

See CommentNotCreatedByCallerException

commit_does_not_exist_exception

See CommitDoesNotExistException

commit_id_does_not_exist_exception

See CommitIdDoesNotExistException

commit_id_required_exception

See CommitIdRequiredException

commit_ids_limit_exceeded_exception

See CommitIdsLimitExceededException

commit_ids_list_required_exception

See CommitIdsListRequiredException

commit_message_length_exceeded_exception

See CommitMessageLengthExceededException

commit_required_exception

See CommitRequiredException

concurrent_reference_update_exception

See ConcurrentReferenceUpdateException

default_branch_cannot_be_deleted_exception

See DefaultBranchCannotBeDeletedException

directory_name_conflicts_with_file_name_exception

See DirectoryNameConflictsWithFileNameException

encryption_integrity_checks_failed_exception

See EncryptionIntegrityChecksFailedException

encryption_key_access_denied_exception

See EncryptionKeyAccessDeniedException

encryption_key_disabled_exception

See EncryptionKeyDisabledException

encryption_key_not_found_exception

See EncryptionKeyNotFoundException

encryption_key_unavailable_exception

See EncryptionKeyUnavailableException

file_content_and_source_file_specified_exception

See FileContentAndSourceFileSpecifiedException

file_content_required_exception

See FileContentRequiredException

file_content_size_limit_exceeded_exception

See FileContentSizeLimitExceededException

file_does_not_exist_exception

See FileDoesNotExistException

file_entry_required_exception

See FileEntryRequiredException

file_mode_required_exception

See FileModeRequiredException

file_name_conflicts_with_directory_name_exception

See FileNameConflictsWithDirectoryNameException

file_path_conflicts_with_submodule_path_exception

See FilePathConflictsWithSubmodulePathException

file_too_large_exception

See FileTooLargeException

folder_content_size_limit_exceeded_exception

See FolderContentSizeLimitExceededException

folder_does_not_exist_exception

See FolderDoesNotExistException

idempotency_parameter_mismatch_exception

See IdempotencyParameterMismatchException

invalid_actor_arn_exception

See InvalidActorArnException

invalid_approval_rule_content_exception

See InvalidApprovalRuleContentException

invalid_approval_rule_name_exception

See InvalidApprovalRuleNameException

invalid_approval_rule_template_content_exception

See InvalidApprovalRuleTemplateContentException

invalid_approval_rule_template_description_exception

See InvalidApprovalRuleTemplateDescriptionException

invalid_approval_rule_template_name_exception

See InvalidApprovalRuleTemplateNameException

invalid_approval_state_exception

See InvalidApprovalStateException

invalid_author_arn_exception

See InvalidAuthorArnException

invalid_blob_id_exception

See InvalidBlobIdException

invalid_branch_name_exception

See InvalidBranchNameException

invalid_client_request_token_exception

See InvalidClientRequestTokenException

invalid_comment_id_exception

See InvalidCommentIdException

invalid_commit_exception

See InvalidCommitException

invalid_commit_id_exception

See InvalidCommitIdException

invalid_conflict_detail_level_exception

See InvalidConflictDetailLevelException

invalid_conflict_resolution_exception

See InvalidConflictResolutionException

invalid_conflict_resolution_strategy_exception

See InvalidConflictResolutionStrategyException

invalid_continuation_token_exception

See InvalidContinuationTokenException

invalid_deletion_parameter_exception

See InvalidDeletionParameterException

invalid_description_exception

See InvalidDescriptionException

invalid_destination_commit_specifier_exception

See InvalidDestinationCommitSpecifierException

invalid_email_exception

See InvalidEmailException

invalid_file_location_exception

See InvalidFileLocationException

invalid_file_mode_exception

See InvalidFileModeException

invalid_file_position_exception

See InvalidFilePositionException

invalid_max_conflict_files_exception

See InvalidMaxConflictFilesException

invalid_max_merge_hunks_exception

See InvalidMaxMergeHunksException

invalid_max_results_exception

See InvalidMaxResultsException

invalid_merge_option_exception

See InvalidMergeOptionException

invalid_order_exception

See InvalidOrderException

invalid_override_status_exception

See InvalidOverrideStatusException

invalid_parent_commit_id_exception

See InvalidParentCommitIdException

invalid_path_exception

See InvalidPathException

invalid_pull_request_event_type_exception

See InvalidPullRequestEventTypeException

invalid_pull_request_id_exception

See InvalidPullRequestIdException

invalid_pull_request_status_exception

See InvalidPullRequestStatusException

invalid_pull_request_status_update_exception

See InvalidPullRequestStatusUpdateException

invalid_reaction_user_arn_exception

See InvalidReactionUserArnException

invalid_reaction_value_exception

See InvalidReactionValueException

invalid_reference_name_exception

See InvalidReferenceNameException

invalid_relative_file_version_enum_exception

See InvalidRelativeFileVersionEnumException

invalid_replacement_content_exception

See InvalidReplacementContentException

invalid_replacement_type_exception

See InvalidReplacementTypeException

invalid_repository_description_exception

See InvalidRepositoryDescriptionException

invalid_repository_name_exception

See InvalidRepositoryNameException

invalid_repository_trigger_branch_name_exception

See InvalidRepositoryTriggerBranchNameException

invalid_repository_trigger_custom_data_exception

See InvalidRepositoryTriggerCustomDataException

invalid_repository_trigger_destination_arn_exception

See InvalidRepositoryTriggerDestinationArnException

invalid_repository_trigger_events_exception

See InvalidRepositoryTriggerEventsException

invalid_repository_trigger_name_exception

See InvalidRepositoryTriggerNameException

invalid_repository_trigger_region_exception

See InvalidRepositoryTriggerRegionException

invalid_resource_arn_exception

See InvalidResourceArnException

invalid_revision_id_exception

See InvalidRevisionIdException

invalid_rule_content_sha256_exception

See InvalidRuleContentSha256Exception

invalid_sort_by_exception

See InvalidSortByException

invalid_source_commit_specifier_exception

See InvalidSourceCommitSpecifierException

invalid_system_tag_usage_exception

See InvalidSystemTagUsageException

invalid_tag_keys_list_exception

See InvalidTagKeysListException

invalid_tags_map_exception

See InvalidTagsMapException

invalid_target_branch_exception

See InvalidTargetBranchException

invalid_target_exception

See InvalidTargetException

invalid_targets_exception

See InvalidTargetsException

invalid_title_exception

See InvalidTitleException

manual_merge_required_exception

See ManualMergeRequiredException

maximum_branches_exceeded_exception

See MaximumBranchesExceededException

maximum_conflict_resolution_entries_exceeded_exception

See MaximumConflictResolutionEntriesExceededException

maximum_file_content_to_load_exceeded_exception

See MaximumFileContentToLoadExceededException

maximum_file_entries_exceeded_exception

See MaximumFileEntriesExceededException

maximum_items_to_compare_exceeded_exception

See MaximumItemsToCompareExceededException

maximum_number_of_approvals_exceeded_exception

See MaximumNumberOfApprovalsExceededException

maximum_open_pull_requests_exceeded_exception

See MaximumOpenPullRequestsExceededException

maximum_repository_names_exceeded_exception

See MaximumRepositoryNamesExceededException

maximum_repository_triggers_exceeded_exception

See MaximumRepositoryTriggersExceededException

maximum_rule_templates_associated_with_repository_exception

See MaximumRuleTemplatesAssociatedWithRepositoryException

merge_option_required_exception

See MergeOptionRequiredException

multiple_conflict_resolution_entries_exception

See MultipleConflictResolutionEntriesException

multiple_repositories_in_pull_request_exception

See MultipleRepositoriesInPullRequestException

name_length_exceeded_exception

See NameLengthExceededException

no_change_exception

See NoChangeException

number_of_rule_templates_exceeded_exception

See NumberOfRuleTemplatesExceededException

number_of_rules_exceeded_exception

See NumberOfRulesExceededException

override_already_set_exception

See OverrideAlreadySetException

override_status_required_exception

See OverrideStatusRequiredException

parent_commit_does_not_exist_exception

See ParentCommitDoesNotExistException

parent_commit_id_outdated_exception

See ParentCommitIdOutdatedException

parent_commit_id_required_exception

See ParentCommitIdRequiredException

path_does_not_exist_exception

See PathDoesNotExistException

path_required_exception

See PathRequiredException

pull_request_already_closed_exception

See PullRequestAlreadyClosedException

pull_request_approval_rules_not_satisfied_exception

See PullRequestApprovalRulesNotSatisfiedException

pull_request_cannot_be_approved_by_author_exception

See PullRequestCannotBeApprovedByAuthorException

pull_request_does_not_exist_exception

See PullRequestDoesNotExistException

pull_request_id_required_exception

See PullRequestIdRequiredException

pull_request_status_required_exception

See PullRequestStatusRequiredException

put_file_entry_conflict_exception

See PutFileEntryConflictException

reaction_limit_exceeded_exception

See ReactionLimitExceededException

reaction_value_required_exception

See ReactionValueRequiredException

reference_does_not_exist_exception

See ReferenceDoesNotExistException

reference_name_required_exception

See ReferenceNameRequiredException

reference_type_not_supported_exception

See ReferenceTypeNotSupportedException

replacement_content_required_exception

See ReplacementContentRequiredException

replacement_type_required_exception

See ReplacementTypeRequiredException

repository_does_not_exist_exception

See RepositoryDoesNotExistException

repository_limit_exceeded_exception

See RepositoryLimitExceededException

repository_name_exists_exception

See RepositoryNameExistsException

repository_name_required_exception

See RepositoryNameRequiredException

repository_names_required_exception

See RepositoryNamesRequiredException

repository_not_associated_with_pull_request_exception

See RepositoryNotAssociatedWithPullRequestException

repository_trigger_branch_name_list_required_exception

See RepositoryTriggerBranchNameListRequiredException

repository_trigger_destination_arn_required_exception

See RepositoryTriggerDestinationArnRequiredException

repository_trigger_events_list_required_exception

See RepositoryTriggerEventsListRequiredException

repository_trigger_name_required_exception

See RepositoryTriggerNameRequiredException

repository_triggers_list_required_exception

See RepositoryTriggersListRequiredException

resource_arn_required_exception

See ResourceArnRequiredException

restricted_source_file_exception

See RestrictedSourceFileException

revision_id_required_exception

See RevisionIdRequiredException

revision_not_current_exception

See RevisionNotCurrentException

same_file_content_exception

See SameFileContentException

same_path_request_exception

See SamePathRequestException

source_and_destination_are_same_exception

See SourceAndDestinationAreSameException

source_file_or_content_required_exception

See SourceFileOrContentRequiredException

tag_keys_list_required_exception

See TagKeysListRequiredException

tag_policy_exception

See TagPolicyException

tags_map_required_exception

See TagsMapRequiredException

target_required_exception

See TargetRequiredException

targets_required_exception

See TargetsRequiredException

tip_of_source_reference_is_different_exception

See TipOfSourceReferenceIsDifferentException

tips_divergence_exceeded_exception

See TipsDivergenceExceededException

title_required_exception

See TitleRequiredException

too_many_tags_exception

See TooManyTagsException

Structs

ActorDoesNotExistException

The specified Amazon Resource Name (ARN) does not exist in the AWS account.

ApprovalRuleContentRequiredException

The content for the approval rule is empty. You must provide some content for an approval rule. The content cannot be null.

ApprovalRuleDoesNotExistException

The specified approval rule does not exist.

ApprovalRuleNameAlreadyExistsException

An approval rule with that name already exists. Approval rule names must be unique

ApprovalRuleNameRequiredException

An approval rule name is required, but was not specified.

ApprovalRuleTemplateContentRequiredException

The content for the approval rule template is empty. You must provide some content for an approval rule template. The content cannot be null.

ApprovalRuleTemplateDoesNotExistException

The specified approval rule template does not exist. Verify that the name is correct and that you are signed in to the AWS Region where the template

ApprovalRuleTemplateInUseException

The approval rule template is associated with one or more repositories. You cannot delete a template that is associated with a repository. Remove

ApprovalRuleTemplateNameAlreadyExistsException

You cannot create an approval rule template with that name because a template with

ApprovalRuleTemplateNameRequiredException

An approval rule template name is required, but was not specified.

ApprovalStateRequiredException

An approval state is required, but was not specified.

AssociateApprovalRuleTemplateWithRepositoryError

AuthorDoesNotExistException

The specified Amazon Resource Name (ARN) does not exist in the AWS account.

BatchAssociateApprovalRuleTemplateWithRepositoriesError

BatchDescribeMergeConflictsError

BatchDisassociateApprovalRuleTemplateFromRepositoriesError

BatchGetCommitsError

BatchGetRepositoriesError

BeforeCommitIdAndAfterCommitIdAreSameException

The before commit ID and the after commit ID are the same, which is not valid. The before commit ID and the after commit ID must be different commit IDs.

BlobIdDoesNotExistException

The specified blob does not exist.

BlobIdRequiredException

A blob ID is required, but was not specified.

BranchDoesNotExistException

The specified branch does not exist.

BranchNameExistsException

Cannot create the branch with the specified name because the commit conflicts with an existing branch with the same name.

BranchNameIsTagNameException

The specified branch name is not valid because it is a tag name. Enter the name of a

BranchNameRequiredException

A branch name is required, but was not specified.

CannotDeleteApprovalRuleFromTemplateException

The approval rule cannot be deleted from the pull request because it was created by an

CannotModifyApprovalRuleFromTemplateException

The approval rule cannot be modified for the pull request because it was created by an

ClientRequestTokenRequiredException

A client request token is required. A client request token is an unique,

CommentContentRequiredException

The comment is empty. You must provide some content for a comment. The content cannot be null.

CommentContentSizeLimitExceededException

The comment is too large. Comments are limited to 1,000 characters.

CommentDeletedException

This comment has already been deleted. You cannot edit or delete a deleted comment.

CommentDoesNotExistException

No comment exists with the provided ID. Verify that you have used the correct ID, and

CommentIdRequiredException

The comment ID is missing or null. A comment ID is required.

CommentNotCreatedByCallerException

You cannot modify or delete this comment. Only comment authors can modify or delete their comments.

CommitDoesNotExistException

The specified commit does not exist or no commit was specified, and the specified repository has no default branch.

CommitIdDoesNotExistException

The specified commit ID does not exist.

CommitIdRequiredException

A commit ID was not specified.

CommitIdsLimitExceededException

The maximum number of allowed commit IDs in a batch request is 100. Verify that your batch requests contains no more than 100 commit IDs, and then try again.

CommitIdsListRequiredException

A list of commit IDs is required, but was either not specified or the list was empty.

CommitMessageLengthExceededException

The commit message is too long. Provide a shorter string.

CommitRequiredException

A commit was not specified.

ConcurrentReferenceUpdateException

The merge cannot be completed because the target branch has been modified. Another user might have modified the target branch while the merge was in progress. Wait a few minutes, and then try again.

CreateApprovalRuleTemplateError

CreateBranchError

CreateCommitError

CreatePullRequestApprovalRuleError

CreatePullRequestError

CreateRepositoryError

CreateUnreferencedMergeCommitError

DefaultBranchCannotBeDeletedException

The specified branch is the default branch for the repository, and cannot be deleted. To delete this branch, you must first set another branch as the default branch.

DeleteApprovalRuleTemplateError

DeleteBranchError

DeleteCommentContentError

DeleteFileError

DeletePullRequestApprovalRuleError

DeleteRepositoryError

DescribeMergeConflictsError

DescribePullRequestEventsError

DirectoryNameConflictsWithFileNameException

A file cannot be added to the repository because the specified path name has the same name as a file that already exists in this repository.

DisassociateApprovalRuleTemplateFromRepositoryError

EncryptionIntegrityChecksFailedException

An encryption integrity check failed.

EncryptionKeyAccessDeniedException

An encryption key could not be accessed.

EncryptionKeyDisabledException

The encryption key is disabled.

EncryptionKeyNotFoundException

No encryption key was found.

EncryptionKeyUnavailableException

The encryption key is not available.

EvaluatePullRequestApprovalRulesError

FileContentAndSourceFileSpecifiedException

The commit cannot be created because both a source file and file content have been

FileContentRequiredException

The file cannot be added because it is empty. Empty files cannot be added to the repository with this API.

FileContentSizeLimitExceededException

The file cannot be added because it is too large. The maximum file size is 6 MB, and

FileDoesNotExistException

The specified file does not exist. Verify that you have used the correct file name,

FileEntryRequiredException

The commit cannot be created because no files have been specified as added, updated, or changed (PutFile or DeleteFile) for the commit.

FileModeRequiredException

The commit cannot be created because no file mode has been specified. A file mode is

FileNameConflictsWithDirectoryNameException

A file cannot be added to the repository because the specified file name has the same name as a directory in this repository. Either provide

FilePathConflictsWithSubmodulePathException

The commit cannot be created because a specified file path points to a submodule. Verify that the destination files

FileTooLargeException

The specified file exceeds the file size limit for AWS CodeCommit. For more information about limits in AWS CodeCommit, see

FolderContentSizeLimitExceededException

The commit cannot be created because at least one of the overall changes in the commit results in a

FolderDoesNotExistException

The specified folder does not exist. Either the folder name is not correct, or you did

GetApprovalRuleTemplateError

GetBlobError

GetBranchError

GetCommentError

GetCommentReactionsError

GetCommentsForComparedCommitError

GetCommentsForPullRequestError

GetCommitError

GetDifferencesError

GetFileError

GetFolderError

GetMergeCommitError

GetMergeConflictsError

GetMergeOptionsError

GetPullRequestApprovalStatesError

GetPullRequestError

GetPullRequestOverrideStateError

GetRepositoryError

GetRepositoryTriggersError

IdempotencyParameterMismatchException

The client request token is not valid. Either the token is not in a valid format, or

InvalidActorArnException

The Amazon Resource Name (ARN) is not valid. Make sure that you have provided the full ARN for the user who initiated the change for the pull request,

InvalidApprovalRuleContentException

The content for the approval rule is not valid.

InvalidApprovalRuleNameException

The name for the approval rule is not valid.

InvalidApprovalRuleTemplateContentException

The content of the approval rule template is not valid.

InvalidApprovalRuleTemplateDescriptionException

The description for the approval rule template is not valid because it exceeds the

InvalidApprovalRuleTemplateNameException

The name of the approval rule template is not valid. Template names must be between 1

InvalidApprovalStateException

The state for the approval is not valid. Valid values include APPROVE and REVOKE.

InvalidAuthorArnException

The Amazon Resource Name (ARN) is not valid. Make sure that you have provided the full ARN for the author of the pull request, and then try again.

InvalidBlobIdException

The specified blob is not valid.

InvalidBranchNameException

The specified reference name is not valid.

InvalidClientRequestTokenException

The client request token is not valid.

InvalidCommentIdException

The comment ID is not in a valid format. Make sure that you have provided the full comment ID.

InvalidCommitException

The specified commit is not valid.

InvalidCommitIdException

The specified commit ID is not valid.

InvalidConflictDetailLevelException

The specified conflict detail level is not valid.

InvalidConflictResolutionException

The specified conflict resolution list is not valid.

InvalidConflictResolutionStrategyException

The specified conflict resolution strategy is not valid.

InvalidContinuationTokenException

The specified continuation token is not valid.

InvalidDeletionParameterException

The specified deletion parameter is not valid.

InvalidDescriptionException

The pull request description is not valid. Descriptions cannot be more than 1,000

InvalidDestinationCommitSpecifierException

The destination commit specifier is not valid. You must provide a valid branch name, tag, or full commit ID.

InvalidEmailException

The specified email address either contains one or more characters that are not allowed, or it exceeds the maximum number of characters

InvalidFileLocationException

The location of the file is not valid. Make sure that you include the file name and

InvalidFileModeException

The specified file mode permission is not valid. For a list of valid file mode permissions, see PutFile.

InvalidFilePositionException

The position is not valid. Make sure that the line number exists in the version of the file you want to comment on.

InvalidMaxConflictFilesException

The specified value for the number of conflict files to return is not valid.

InvalidMaxMergeHunksException

The specified value for the number of merge hunks to return is not valid.

InvalidMaxResultsException

The specified number of maximum results is not valid.

InvalidMergeOptionException

The specified merge option is not valid for this operation. Not all merge strategies are supported for all operations.

InvalidOrderException

The specified sort order is not valid.

InvalidOverrideStatusException

The override status is not valid. Valid statuses are OVERRIDE and REVOKE.

InvalidParentCommitIdException

The parent commit ID is not valid. The commit ID cannot be empty, and must match the head commit ID for the branch of the repository where you

InvalidPathException

The specified path is not valid.

InvalidPullRequestEventTypeException

The pull request event type is not valid.

InvalidPullRequestIdException

The pull request ID is not valid. Make sure that you have provided the full ID and that the pull request is in the specified repository, and then try again.

InvalidPullRequestStatusException

The pull request status is not valid. The only valid values are OPEN and CLOSED.

InvalidPullRequestStatusUpdateException

The pull request status update is not valid. The only valid update is from OPEN to CLOSED.

InvalidReactionUserArnException

The Amazon Resource Name (ARN) of the user or identity is not valid.

InvalidReactionValueException

The value of the reaction is not valid. For more information, see the AWS CodeCommit User Guide.

InvalidReferenceNameException

The specified reference name format is not valid. Reference names must conform to the

InvalidRelativeFileVersionEnumException

Either the enum is not in a valid format, or the specified file version enum is not valid in respect to the current file version.

InvalidReplacementContentException

Automerge was specified for resolving the conflict, but the replacement type is not valid or content is missing.

InvalidReplacementTypeException

Automerge was specified for resolving the conflict, but the specified replacement type is not valid.

InvalidRepositoryDescriptionException

The specified repository description is not valid.

InvalidRepositoryNameException

A specified repository name is not valid.

InvalidRepositoryTriggerBranchNameException

One or more branch names specified for the trigger is not valid.

InvalidRepositoryTriggerCustomDataException

The custom data provided for the trigger is not valid.

InvalidRepositoryTriggerDestinationArnException

The Amazon Resource Name (ARN) for the trigger is not valid for the specified destination. The most common reason for this error is that the ARN does not meet the requirements for the service type.

InvalidRepositoryTriggerEventsException

One or more events specified for the trigger is not valid. Check to make sure that all events specified match the requirements for allowed events.

InvalidRepositoryTriggerNameException

The name of the trigger is not valid.

InvalidRepositoryTriggerRegionException

The AWS Region for the trigger target does not match the AWS Region for the

InvalidResourceArnException

The value for the resource ARN is not valid. For more information about resources in AWS CodeCommit, see

InvalidRevisionIdException

The revision ID is not valid. Use GetPullRequest to determine the value.

InvalidRuleContentSha256Exception

The SHA-256 hash signature for the rule content is not valid.

InvalidSortByException

The specified sort by value is not valid.

InvalidSourceCommitSpecifierException

The source commit specifier is not valid. You must provide a valid branch name, tag, or full commit ID.

InvalidSystemTagUsageException

The specified tag is not valid. Key names cannot be prefixed with aws:.

InvalidTagKeysListException

The list of tags is not valid.

InvalidTagsMapException

The map of tags is not valid.

InvalidTargetBranchException

The specified target branch is not valid.

InvalidTargetException

The target for the pull request is not valid. A target must contain the full values for the repository name, source branch, and destination branch for the pull request.

InvalidTargetsException

The targets for the pull request is not valid or not in a valid format. Targets are a list of target objects. Each target object must contain the full values for

InvalidTitleException

The title of the pull request is not valid. Pull request titles cannot exceed 100 characters in length.

ListApprovalRuleTemplatesError

ListAssociatedApprovalRuleTemplatesForRepositoryError

ListBranchesError

ListPullRequestsError

ListRepositoriesError

ListRepositoriesForApprovalRuleTemplateError

ListTagsForResourceError

ManualMergeRequiredException

The pull request cannot be merged automatically into the destination branch. You must manually merge the branches and resolve any conflicts.

MaximumBranchesExceededException

The number of branches for the trigger was exceeded.

MaximumConflictResolutionEntriesExceededException

The number of allowed conflict resolution entries was exceeded.

MaximumFileContentToLoadExceededException

The number of files to load exceeds the allowed limit.

MaximumFileEntriesExceededException

The number of specified files to change as part of this commit exceeds the maximum number of files

MaximumItemsToCompareExceededException

The number of items to compare between the source or destination branches and the merge base has exceeded the maximum allowed.

MaximumNumberOfApprovalsExceededException

The number of approvals required for the approval rule exceeds the maximum number allowed.

MaximumOpenPullRequestsExceededException

You cannot create the pull request because the repository has too many open pull requests.

MaximumRepositoryNamesExceededException

The maximum number of allowed repository names was exceeded. Currently, this number is 100.

MaximumRepositoryTriggersExceededException

The number of triggers allowed for the repository was exceeded.

MaximumRuleTemplatesAssociatedWithRepositoryException

The maximum number of approval rule templates for a repository has been exceeded. You cannot associate more than 25

MergeBranchesByFastForwardError

MergeBranchesBySquashError

MergeBranchesByThreeWayError

MergeOptionRequiredException

A merge option or stategy is required, and none was provided.

MergePullRequestByFastForwardError

MergePullRequestBySquashError

MergePullRequestByThreeWayError

MultipleConflictResolutionEntriesException

More than one conflict resolution entries exists for the conflict. A conflict can have only one conflict resolution entry.

MultipleRepositoriesInPullRequestException

You cannot include more than one repository in a pull request. Make sure you have specified only one repository name in your request, and then try again.

NameLengthExceededException

The user name is not valid because it has exceeded the character limit for author names.

NoChangeException

The commit cannot be created because no changes will be made to the repository as a result of this commit. A commit must contain at least one change.

NumberOfRuleTemplatesExceededException

The maximum number of approval rule templates has been exceeded for this AWS Region.

NumberOfRulesExceededException

The approval rule cannot be added. The pull request has the maximum number of approval rules associated with it.

OverrideAlreadySetException

The pull request has already had its approval rules set to override.

OverridePullRequestApprovalRulesError

OverrideStatusRequiredException

An override status is required, but no value was provided. Valid values include OVERRIDE and REVOKE.

ParentCommitDoesNotExistException

The parent commit ID is not valid because it does not exist. The specified parent commit ID does not exist in the specified branch of the repository.

ParentCommitIdOutdatedException

The file could not be added because the provided parent commit ID is not the current tip of the specified branch. To view the full commit ID of the current head

ParentCommitIdRequiredException

A parent commit ID is required. To view the full commit ID of a branch in a repository, use GetBranch or a Git command

PathDoesNotExistException

The specified path does not exist.

PathRequiredException

The folderPath for a location cannot be null.

PostCommentForComparedCommitError

PostCommentForPullRequestError

PostCommentReplyError

PullRequestAlreadyClosedException

The pull request status cannot be updated because it is already closed.

PullRequestApprovalRulesNotSatisfiedException

The pull request cannot be merged because one or more approval rules applied to the pull request have conditions that have not been met.

PullRequestCannotBeApprovedByAuthorException

The approval cannot be applied because the user approving the pull request matches the user who created the pull request. You cannot approve a pull

PullRequestDoesNotExistException

The pull request ID could not be found. Make sure that you have specified the correct repository name and pull request ID, and then try again.

PullRequestIdRequiredException

A pull request ID is required, but none was provided.

PullRequestStatusRequiredException

A pull request status is required, but none was provided.

PutCommentReactionError

PutFileEntryConflictException

The commit cannot be created because one or more files specified in the commit reference both a file and a folder.

PutFileError

PutRepositoryTriggersError

ReactionLimitExceededException

The number of reactions has been exceeded. Reactions are limited to one reaction per user for each individual comment ID.

ReactionValueRequiredException

A reaction value is required.

ReferenceDoesNotExistException

The specified reference does not exist. You must provide a full commit ID.

ReferenceNameRequiredException

A reference name is required, but none was provided.

ReferenceTypeNotSupportedException

The specified reference is not a supported type.

ReplacementContentRequiredException

USE_NEW_CONTENT was specified, but no replacement content has been provided.

ReplacementTypeRequiredException

A replacement type is required.

RepositoryDoesNotExistException

The specified repository does not exist.

RepositoryLimitExceededException

A repository resource limit was exceeded.

RepositoryNameExistsException

The specified repository name already exists.

RepositoryNameRequiredException

A repository name is required, but was not specified.

RepositoryNamesRequiredException

At least one repository name object is required, but was not specified.

RepositoryNotAssociatedWithPullRequestException

The repository does not contain any pull requests with that pull request ID. Use GetPullRequest to verify the correct repository name for the pull request ID.

RepositoryTriggerBranchNameListRequiredException

At least one branch name is required, but was not specified in the trigger

RepositoryTriggerDestinationArnRequiredException

A destination ARN for the target service for the trigger is required, but was not

RepositoryTriggerEventsListRequiredException

At least one event for the trigger is required, but was not specified.

RepositoryTriggerNameRequiredException

A name for the trigger is required, but was not specified.

RepositoryTriggersListRequiredException

The list of triggers for the repository is required, but was not specified.

ResourceArnRequiredException

A valid Amazon Resource Name (ARN) for an AWS CodeCommit resource is required. For a list of valid resources in AWS CodeCommit, see

RestrictedSourceFileException

The commit cannot be created because one of the changes specifies copying or moving a .gitkeep file.

RevisionIdRequiredException

A revision ID is required, but was not provided.

RevisionNotCurrentException

The revision ID provided in the request does not match the current revision ID. Use GetPullRequest to retrieve the current revision ID.

SameFileContentException

The file was not added or updated because the content of the file is exactly the same as the content of that file in the repository and branch

SamePathRequestException

The commit cannot be created because one or more changes in this commit duplicate actions in the same file path. For example,

SourceAndDestinationAreSameException

The source branch and destination branch for the pull request are the same. You must

SourceFileOrContentRequiredException

The commit cannot be created because no source files or file content have been specified for the commit.

TagKeysListRequiredException

A list of tag keys is required. The list cannot be empty or null.

TagPolicyException

The tag policy is not valid.

TagResourceError

TagsMapRequiredException

A map of tags is required.

TargetRequiredException

A pull request target is required. It cannot be empty or null. A pull request target must contain the full values for the repository name, source branch, and destination branch for the pull request.

TargetsRequiredException

An array of target objects is required. It cannot be empty or null.

TestRepositoryTriggersError

TipOfSourceReferenceIsDifferentException

The tip of the source branch in the destination repository does not match the tip of the source branch specified in your request.

TipsDivergenceExceededException

The divergence between the tips of the provided commit specifiers is too great to determine whether there might be

TitleRequiredException

A pull request title is required. It cannot be empty or null.

TooManyTagsException

The maximum number of tags for an AWS CodeCommit resource has been exceeded.

UntagResourceError

UpdateApprovalRuleTemplateContentError

UpdateApprovalRuleTemplateDescriptionError

UpdateApprovalRuleTemplateNameError

UpdateCommentError

UpdateDefaultBranchError

UpdatePullRequestApprovalRuleContentError

UpdatePullRequestApprovalStateError

UpdatePullRequestDescriptionError

UpdatePullRequestStatusError

UpdatePullRequestTitleError

UpdateRepositoryDescriptionError

UpdateRepositoryNameError

Enums

AssociateApprovalRuleTemplateWithRepositoryErrorKind

BatchAssociateApprovalRuleTemplateWithRepositoriesErrorKind

BatchDescribeMergeConflictsErrorKind

BatchDisassociateApprovalRuleTemplateFromRepositoriesErrorKind

BatchGetCommitsErrorKind

BatchGetRepositoriesErrorKind

CreateApprovalRuleTemplateErrorKind

CreateBranchErrorKind

CreateCommitErrorKind

CreatePullRequestApprovalRuleErrorKind

CreatePullRequestErrorKind

CreateRepositoryErrorKind

CreateUnreferencedMergeCommitErrorKind

DeleteApprovalRuleTemplateErrorKind

DeleteBranchErrorKind

DeleteCommentContentErrorKind

DeleteFileErrorKind

DeletePullRequestApprovalRuleErrorKind

DeleteRepositoryErrorKind

DescribeMergeConflictsErrorKind

DescribePullRequestEventsErrorKind

DisassociateApprovalRuleTemplateFromRepositoryErrorKind

EvaluatePullRequestApprovalRulesErrorKind

GetApprovalRuleTemplateErrorKind

GetBlobErrorKind

GetBranchErrorKind

GetCommentErrorKind

GetCommentReactionsErrorKind

GetCommentsForComparedCommitErrorKind

GetCommentsForPullRequestErrorKind

GetCommitErrorKind

GetDifferencesErrorKind

GetFileErrorKind

GetFolderErrorKind

GetMergeCommitErrorKind

GetMergeConflictsErrorKind

GetMergeOptionsErrorKind

GetPullRequestApprovalStatesErrorKind

GetPullRequestErrorKind

GetPullRequestOverrideStateErrorKind

GetRepositoryErrorKind

GetRepositoryTriggersErrorKind

ListApprovalRuleTemplatesErrorKind

ListAssociatedApprovalRuleTemplatesForRepositoryErrorKind

ListBranchesErrorKind

ListPullRequestsErrorKind

ListRepositoriesErrorKind

ListRepositoriesForApprovalRuleTemplateErrorKind

ListTagsForResourceErrorKind

MergeBranchesByFastForwardErrorKind

MergeBranchesBySquashErrorKind

MergeBranchesByThreeWayErrorKind

MergePullRequestByFastForwardErrorKind

MergePullRequestBySquashErrorKind

MergePullRequestByThreeWayErrorKind

OverridePullRequestApprovalRulesErrorKind

PostCommentForComparedCommitErrorKind

PostCommentForPullRequestErrorKind

PostCommentReplyErrorKind

PutCommentReactionErrorKind

PutFileErrorKind

PutRepositoryTriggersErrorKind

TagResourceErrorKind

TestRepositoryTriggersErrorKind

UntagResourceErrorKind

UpdateApprovalRuleTemplateContentErrorKind

UpdateApprovalRuleTemplateDescriptionErrorKind

UpdateApprovalRuleTemplateNameErrorKind

UpdateCommentErrorKind

UpdateDefaultBranchErrorKind

UpdatePullRequestApprovalRuleContentErrorKind

UpdatePullRequestApprovalStateErrorKind

UpdatePullRequestDescriptionErrorKind

UpdatePullRequestStatusErrorKind

UpdatePullRequestTitleErrorKind

UpdateRepositoryDescriptionErrorKind

UpdateRepositoryNameErrorKind