Module aws_sdk_codecommit::input

Modules

associate_approval_rule_template_with_repository_input

See AssociateApprovalRuleTemplateWithRepositoryInput

batch_associate_approval_rule_template_with_repositories_input

See BatchAssociateApprovalRuleTemplateWithRepositoriesInput

batch_describe_merge_conflicts_input

See BatchDescribeMergeConflictsInput

batch_disassociate_approval_rule_template_from_repositories_input

See BatchDisassociateApprovalRuleTemplateFromRepositoriesInput

batch_get_commits_input

See BatchGetCommitsInput

batch_get_repositories_input

See BatchGetRepositoriesInput

create_approval_rule_template_input

See CreateApprovalRuleTemplateInput

create_branch_input

See CreateBranchInput

create_commit_input

See CreateCommitInput

create_pull_request_approval_rule_input

See CreatePullRequestApprovalRuleInput

create_pull_request_input

See CreatePullRequestInput

create_repository_input

See CreateRepositoryInput

create_unreferenced_merge_commit_input

See CreateUnreferencedMergeCommitInput

delete_approval_rule_template_input

See DeleteApprovalRuleTemplateInput

delete_branch_input

See DeleteBranchInput

delete_comment_content_input

See DeleteCommentContentInput

delete_file_input

See DeleteFileInput

delete_pull_request_approval_rule_input

See DeletePullRequestApprovalRuleInput

delete_repository_input

See DeleteRepositoryInput

describe_merge_conflicts_input

See DescribeMergeConflictsInput

describe_pull_request_events_input

See DescribePullRequestEventsInput

disassociate_approval_rule_template_from_repository_input

See DisassociateApprovalRuleTemplateFromRepositoryInput

evaluate_pull_request_approval_rules_input

See EvaluatePullRequestApprovalRulesInput

get_approval_rule_template_input

See GetApprovalRuleTemplateInput

get_blob_input

See GetBlobInput

get_branch_input

See GetBranchInput

get_comment_input

See GetCommentInput

get_comment_reactions_input

See GetCommentReactionsInput

get_comments_for_compared_commit_input

See GetCommentsForComparedCommitInput

get_comments_for_pull_request_input

See GetCommentsForPullRequestInput

get_commit_input

See GetCommitInput

get_differences_input

See GetDifferencesInput

get_file_input

See GetFileInput

get_folder_input

See GetFolderInput

get_merge_commit_input

See GetMergeCommitInput

get_merge_conflicts_input

See GetMergeConflictsInput

get_merge_options_input

See GetMergeOptionsInput

get_pull_request_approval_states_input

See GetPullRequestApprovalStatesInput

get_pull_request_input

See GetPullRequestInput

get_pull_request_override_state_input

See GetPullRequestOverrideStateInput

get_repository_input

See GetRepositoryInput

get_repository_triggers_input

See GetRepositoryTriggersInput

list_approval_rule_templates_input

See ListApprovalRuleTemplatesInput

list_associated_approval_rule_templates_for_repository_input

See ListAssociatedApprovalRuleTemplatesForRepositoryInput

list_branches_input

See ListBranchesInput

list_pull_requests_input

See ListPullRequestsInput

list_repositories_for_approval_rule_template_input

See ListRepositoriesForApprovalRuleTemplateInput

list_repositories_input

See ListRepositoriesInput

list_tags_for_resource_input

See ListTagsForResourceInput

merge_branches_by_fast_forward_input

See MergeBranchesByFastForwardInput

merge_branches_by_squash_input

See MergeBranchesBySquashInput

merge_branches_by_three_way_input

See MergeBranchesByThreeWayInput

merge_pull_request_by_fast_forward_input

See MergePullRequestByFastForwardInput

merge_pull_request_by_squash_input

See MergePullRequestBySquashInput

merge_pull_request_by_three_way_input

See MergePullRequestByThreeWayInput

override_pull_request_approval_rules_input

See OverridePullRequestApprovalRulesInput

post_comment_for_compared_commit_input

See PostCommentForComparedCommitInput

post_comment_for_pull_request_input

See PostCommentForPullRequestInput

post_comment_reply_input

See PostCommentReplyInput

put_comment_reaction_input

See PutCommentReactionInput

put_file_input

See PutFileInput

put_repository_triggers_input

See PutRepositoryTriggersInput

tag_resource_input

See TagResourceInput

test_repository_triggers_input

See TestRepositoryTriggersInput

untag_resource_input

See UntagResourceInput

update_approval_rule_template_content_input

See UpdateApprovalRuleTemplateContentInput

update_approval_rule_template_description_input

See UpdateApprovalRuleTemplateDescriptionInput

update_approval_rule_template_name_input

See UpdateApprovalRuleTemplateNameInput

update_comment_input

See UpdateCommentInput

update_default_branch_input

See UpdateDefaultBranchInput

update_pull_request_approval_rule_content_input

See UpdatePullRequestApprovalRuleContentInput

update_pull_request_approval_state_input

See UpdatePullRequestApprovalStateInput

update_pull_request_description_input

See UpdatePullRequestDescriptionInput

update_pull_request_status_input

See UpdatePullRequestStatusInput

update_pull_request_title_input

See UpdatePullRequestTitleInput

update_repository_description_input

See UpdateRepositoryDescriptionInput

update_repository_name_input

See UpdateRepositoryNameInput

Structs

AssociateApprovalRuleTemplateWithRepositoryInput

BatchAssociateApprovalRuleTemplateWithRepositoriesInput

BatchDescribeMergeConflictsInput

BatchDisassociateApprovalRuleTemplateFromRepositoriesInput

BatchGetCommitsInput

BatchGetRepositoriesInput

Represents the input of a batch get repositories operation.

CreateApprovalRuleTemplateInput

CreateBranchInput

Represents the input of a create branch operation.

CreateCommitInput

CreatePullRequestApprovalRuleInput

CreatePullRequestInput

CreateRepositoryInput

Represents the input of a create repository operation.

CreateUnreferencedMergeCommitInput

DeleteApprovalRuleTemplateInput

DeleteBranchInput

Represents the input of a delete branch operation.

DeleteCommentContentInput

DeleteFileInput

DeletePullRequestApprovalRuleInput

DeleteRepositoryInput

Represents the input of a delete repository operation.

DescribeMergeConflictsInput

DescribePullRequestEventsInput

DisassociateApprovalRuleTemplateFromRepositoryInput

EvaluatePullRequestApprovalRulesInput

GetApprovalRuleTemplateInput

GetBlobInput

Represents the input of a get blob operation.

GetBranchInput

Represents the input of a get branch operation.

GetCommentInput

GetCommentReactionsInput

GetCommentsForComparedCommitInput

GetCommentsForPullRequestInput

GetCommitInput

Represents the input of a get commit operation.

GetDifferencesInput

GetFileInput

GetFolderInput

GetMergeCommitInput

GetMergeConflictsInput

GetMergeOptionsInput

GetPullRequestApprovalStatesInput

GetPullRequestInput

GetPullRequestOverrideStateInput

GetRepositoryInput

Represents the input of a get repository operation.

GetRepositoryTriggersInput

Represents the input of a get repository triggers operation.

ListApprovalRuleTemplatesInput

ListAssociatedApprovalRuleTemplatesForRepositoryInput

ListBranchesInput

Represents the input of a list branches operation.

ListPullRequestsInput

ListRepositoriesForApprovalRuleTemplateInput

ListRepositoriesInput

Represents the input of a list repositories operation.

ListTagsForResourceInput

MergeBranchesByFastForwardInput

MergeBranchesBySquashInput

MergeBranchesByThreeWayInput

MergePullRequestByFastForwardInput

MergePullRequestBySquashInput

MergePullRequestByThreeWayInput

OverridePullRequestApprovalRulesInput

PostCommentForComparedCommitInput

PostCommentForPullRequestInput

PostCommentReplyInput

PutCommentReactionInput

PutFileInput

PutRepositoryTriggersInput

Represents the input of a put repository triggers operation.

TagResourceInput

TestRepositoryTriggersInput

Represents the input of a test repository triggers operation.

UntagResourceInput

UpdateApprovalRuleTemplateContentInput

UpdateApprovalRuleTemplateDescriptionInput

UpdateApprovalRuleTemplateNameInput

UpdateCommentInput

UpdateDefaultBranchInput

Represents the input of an update default branch operation.

UpdatePullRequestApprovalRuleContentInput

UpdatePullRequestApprovalStateInput

UpdatePullRequestDescriptionInput

UpdatePullRequestStatusInput

UpdatePullRequestTitleInput

UpdateRepositoryDescriptionInput

Represents the input of an update repository description operation.

UpdateRepositoryNameInput

Represents the input of an update repository description operation.