Module aws_sdk_codecommit::model

Modules

approval

See Approval

approval_rule

See ApprovalRule

approval_rule_event_metadata

See ApprovalRuleEventMetadata

approval_rule_overridden_event_metadata

See ApprovalRuleOverriddenEventMetadata

approval_rule_template

See ApprovalRuleTemplate

approval_state_changed_event_metadata

See ApprovalStateChangedEventMetadata

batch_associate_approval_rule_template_with_repositories_error

See BatchAssociateApprovalRuleTemplateWithRepositoriesError

batch_describe_merge_conflicts_error

See BatchDescribeMergeConflictsError

batch_disassociate_approval_rule_template_from_repositories_error

See BatchDisassociateApprovalRuleTemplateFromRepositoriesError

batch_get_commits_error

See BatchGetCommitsError

blob_metadata

See BlobMetadata

branch_info

See BranchInfo

comment

See Comment

comments_for_compared_commit

See CommentsForComparedCommit

comments_for_pull_request

See CommentsForPullRequest

commit

See Commit

conflict

See Conflict

conflict_metadata

See ConflictMetadata

conflict_resolution

See ConflictResolution

delete_file_entry

See DeleteFileEntry

difference

See Difference

evaluation

See Evaluation

file

See File

file_metadata

See FileMetadata

file_modes

See FileModes

file_sizes

See FileSizes

folder

See Folder

is_binary_file

See IsBinaryFile

location

See Location

merge_hunk

See MergeHunk

merge_hunk_detail

See MergeHunkDetail

merge_metadata

See MergeMetadata

merge_operations

See MergeOperations

object_types

See ObjectTypes

origin_approval_rule_template

See OriginApprovalRuleTemplate

pull_request

See PullRequest

pull_request_created_event_metadata

See PullRequestCreatedEventMetadata

pull_request_event

See PullRequestEvent

pull_request_merged_state_changed_event_metadata

See PullRequestMergedStateChangedEventMetadata

pull_request_source_reference_updated_event_metadata

See PullRequestSourceReferenceUpdatedEventMetadata

pull_request_status_changed_event_metadata

See PullRequestStatusChangedEventMetadata

pull_request_target

See PullRequestTarget

put_file_entry

See PutFileEntry

reaction_for_comment

See ReactionForComment

reaction_value_formats

See ReactionValueFormats

replace_content_entry

See ReplaceContentEntry

repository_metadata

See RepositoryMetadata

repository_name_id_pair

See RepositoryNameIdPair

repository_trigger

See RepositoryTrigger

repository_trigger_execution_failure

See RepositoryTriggerExecutionFailure

set_file_mode_entry

See SetFileModeEntry

source_file_specifier

See SourceFileSpecifier

sub_module

See SubModule

symbolic_link

See SymbolicLink

target

See Target

user_info

See UserInfo

Structs

Approval

Returns information about a specific approval on a pull request.

ApprovalRule

Returns information about an approval rule.

ApprovalRuleEventMetadata

Returns information about an event for an approval rule.

ApprovalRuleOverriddenEventMetadata

Returns information about an override event for approval rules for a pull request.

ApprovalRuleTemplate

Returns information about an approval rule template.

ApprovalStateChangedEventMetadata

Returns information about a change in the approval state for a pull request.

BatchAssociateApprovalRuleTemplateWithRepositoriesError

Returns information about errors in a BatchAssociateApprovalRuleTemplateWithRepositories operation.

BatchDescribeMergeConflictsError

Returns information about errors in a BatchDescribeMergeConflicts operation.

BatchDisassociateApprovalRuleTemplateFromRepositoriesError

Returns information about errors in a BatchDisassociateApprovalRuleTemplateFromRepositories operation.

BatchGetCommitsError

Returns information about errors in a BatchGetCommits operation.

BlobMetadata

Returns information about a specific Git blob object.

BranchInfo

Returns information about a branch.

Comment

Returns information about a specific comment.

CommentsForComparedCommit

Returns information about comments on the comparison between two commits.

CommentsForPullRequest

Returns information about comments on a pull request.

Commit

Returns information about a specific commit.

Conflict

Information about conflicts in a merge operation.

ConflictMetadata

Information about the metadata for a conflict in a merge operation.

ConflictResolution

If AUTOMERGE is the conflict resolution strategy, a list of inputs to use when

DeleteFileEntry

A file that is deleted as part of a commit.

Difference

Returns information about a set of differences for a commit specifier.

Evaluation

Returns information about the approval rules applied to a pull request and whether conditions have been met.

File

Returns information about a file in a repository.

FileMetadata

A file to be added, updated, or deleted as part of a commit.

FileModes

Information about file modes in a merge or pull request.

FileSizes

Information about the size of files in a merge or pull request.

Folder

Returns information about a folder in a repository.

IsBinaryFile

Information about whether a file is binary or textual in a merge or pull request operation.

Location

Returns information about the location of a change or comment in the comparison between two commits or a pull request.

MergeHunk

Information about merge hunks in a merge or pull request operation.

MergeHunkDetail

Information about the details of a merge hunk that contains a conflict in a merge or pull request operation.

MergeMetadata

Returns information about a merge or potential merge between a source reference and a destination reference in a pull request.

MergeOperations

Information about the file operation conflicts in a merge operation.

ObjectTypes

Information about the type of an object in a merge operation.

OriginApprovalRuleTemplate

Returns information about the template that created the approval rule for a pull request.

PullRequest

Returns information about a pull request.

PullRequestCreatedEventMetadata

Metadata about the pull request that is used when comparing the pull request source with its destination.

PullRequestEvent

Returns information about a pull request event.

PullRequestMergedStateChangedEventMetadata

Returns information about the change in the merge state for a pull request event.

PullRequestSourceReferenceUpdatedEventMetadata

Information about an update to the source branch of a pull request.

PullRequestStatusChangedEventMetadata

Information about a change to the status of a pull request.

PullRequestTarget

Returns information about a pull request target.

PutFileEntry

Information about a file added or updated as part of a commit.

ReactionForComment

Information about the reaction values provided by users on a comment.

ReactionValueFormats

Information about the values for reactions to a comment. AWS CodeCommit supports a limited set of reactions.

ReplaceContentEntry

Information about a replacement content entry in the conflict of a merge or pull request operation.

RepositoryMetadata

Information about a repository.

RepositoryNameIdPair

Information about a repository name and ID.

RepositoryTrigger

Information about a trigger for a repository.

RepositoryTriggerExecutionFailure

A trigger failed to run.

SetFileModeEntry

Information about the file mode changes.

SourceFileSpecifier

Information about a source file that is part of changes made in a commit.

SubModule

Returns information about a submodule reference in a repository folder.

SymbolicLink

Returns information about a symbolic link in a repository folder.

Target

Returns information about a target for a pull request.

UserInfo

Information about the user who made a specified commit.

Enums

ApprovalState

ChangeTypeEnum

ConflictDetailLevelTypeEnum

ConflictResolutionStrategyTypeEnum

FileModeTypeEnum

MergeOptionTypeEnum

ObjectTypeEnum

OrderEnum

OverrideStatus

PullRequestEventType

PullRequestStatusEnum

RelativeFileVersionEnum

ReplacementTypeEnum

RepositoryTriggerEventEnum

SortByEnum