Module aws_sdk_codecommit::operation

Structs

AssociateApprovalRuleTemplateWithRepository

Creates an association between an approval rule template and a specified repository.

BatchAssociateApprovalRuleTemplateWithRepositories

Creates an association between an approval rule template and one or more specified repositories.

BatchDescribeMergeConflicts

Returns information about one or more merge conflicts in the attempted merge of two commit specifiers using the squash or three-way merge strategy.

BatchDisassociateApprovalRuleTemplateFromRepositories

Removes the association between an approval rule template and one or more specified repositories.

BatchGetCommits

Returns information about the contents of one or more commits in a repository.

BatchGetRepositories

Returns information about one or more repositories.

CreateApprovalRuleTemplate

Creates a template for approval rules that can then be associated with one or more

CreateBranch

Creates a branch in a repository and points the branch to a commit.

CreateCommit

Creates a commit for a repository on the tip of a specified branch.

CreatePullRequest

Creates a pull request in the specified repository.

CreatePullRequestApprovalRule

Creates an approval rule for a pull request.

CreateRepository

Creates a new, empty repository.

CreateUnreferencedMergeCommit

Creates an unreferenced commit that represents the result of merging two branches

DeleteApprovalRuleTemplate

Deletes a specified approval rule template. Deleting a template does not remove approval rules on pull requests already created with the template.

DeleteBranch

Deletes a branch from a repository, unless that branch is the default branch for the repository.

DeleteCommentContent

Deletes the content of a comment made on a change, file, or commit in a repository.

DeleteFile

Deletes a specified file from a specified branch. A commit is created on the branch

DeletePullRequestApprovalRule

Deletes an approval rule from a specified pull request. Approval rules can be deleted from a pull request only if the pull request is open, and if the

DeleteRepository

Deletes a repository. If a specified repository was already deleted, a null repository

DescribeMergeConflicts

Returns information about one or more merge conflicts in the attempted merge of two

DescribePullRequestEvents

Returns information about one or more pull request events.

DisassociateApprovalRuleTemplateFromRepository

Removes the association between a template and a repository so that approval rules

EvaluatePullRequestApprovalRules

Evaluates whether a pull request has met all the conditions specified in its associated approval rules.

GetApprovalRuleTemplate

Returns information about a specified approval rule template.

GetBlob

Returns the base-64 encoded content of an individual blob in a repository.

GetBranch

Returns information about a repository branch, including its name and the last commit ID.

GetComment

Returns the content of a comment made on a change, file, or commit in a repository.

GetCommentReactions

Returns information about reactions to a specified comment ID. Reactions from users who have been deleted will not be included in the count.

GetCommentsForComparedCommit

Returns information about comments made on the comparison between two commits.

GetCommentsForPullRequest

Returns comments made on a pull request.

GetCommit

Returns information about a commit, including commit message and committer information.

GetDifferences

Returns information about the differences in a valid commit specifier (such as a

GetFile

Returns the base-64 encoded contents of a specified file and its metadata.

GetFolder

Returns the contents of a specified folder in a repository.

GetMergeCommit

Returns information about a specified merge commit.

GetMergeConflicts

Returns information about merge conflicts between the before and after commit IDs for a pull request in a repository.

GetMergeOptions

Returns information about the merge options available for merging two specified

GetPullRequest

Gets information about a pull request in a specified repository.

GetPullRequestApprovalStates

Gets information about the approval states for a specified pull request. Approval states only apply to pull requests that have one or more

GetPullRequestOverrideState

Returns information about whether approval rules have been set aside (overridden) for a

GetRepository

Returns information about a repository.

GetRepositoryTriggers

Gets information about triggers configured for a repository.

ListApprovalRuleTemplates

Lists all approval rule templates in the specified AWS Region in your AWS account. If

ListAssociatedApprovalRuleTemplatesForRepository

Lists all approval rule templates that are associated with a specified repository.

ListBranches

Gets information about one or more branches in a repository.

ListPullRequests

Returns a list of pull requests for a specified repository. The return list can be refined by pull request

ListRepositories

Gets information about one or more repositories.

ListRepositoriesForApprovalRuleTemplate

Lists all repositories associated with the specified approval rule template.

ListTagsForResource

Gets information about AWS tags for a specified Amazon Resource Name (ARN) in AWS

MergeBranchesByFastForward

Merges two branches using the fast-forward merge strategy.

MergeBranchesBySquash

Merges two branches using the squash merge strategy.

MergeBranchesByThreeWay

Merges two specified branches using the three-way merge strategy.

MergePullRequestByFastForward

Attempts to merge the source commit of a pull request into the specified destination

MergePullRequestBySquash

Attempts to merge the source commit of a pull request into the specified destination

MergePullRequestByThreeWay

Attempts to merge the source commit of a pull request into the specified destination

OverridePullRequestApprovalRules

Sets aside (overrides) all approval rule requirements for a specified pull request.

PostCommentForComparedCommit

Posts a comment on the comparison between two commits.

PostCommentForPullRequest

Posts a comment on a pull request.

PostCommentReply

Posts a comment in reply to an existing comment on a comparison between commits or a pull request.

PutCommentReaction

Adds or updates a reaction to a specified comment for the user whose identity is used to make the request. You can only add or

PutFile

Adds or updates a file in a branch in an AWS CodeCommit repository, and generates a commit for the addition in the specified branch.

PutRepositoryTriggers

Replaces all triggers for a repository. Used to create or delete triggers.

TagResource

Adds or updates tags for a resource in AWS CodeCommit. For a list of valid resources

TestRepositoryTriggers

Tests the functionality of repository triggers by sending information to the trigger

UntagResource

Removes tags for a resource in AWS CodeCommit. For a list of valid resources in AWS

UpdateApprovalRuleTemplateContent

Updates the content of an approval rule template. You can change the number of

UpdateApprovalRuleTemplateDescription

Updates the description for a specified approval rule template.

UpdateApprovalRuleTemplateName

Updates the name of a specified approval rule template.

UpdateComment

Replaces the contents of a comment.

UpdateDefaultBranch

Sets or changes the default branch name for the specified repository.

UpdatePullRequestApprovalRuleContent

Updates the structure of an approval rule created specifically for a pull request. For example, you can change the number of required approvers and

UpdatePullRequestApprovalState

Updates the state of a user's approval on a pull request. The user is derived from the signed-in account when the request is made.

UpdatePullRequestDescription

Replaces the contents of the description of a pull request.

UpdatePullRequestStatus

Updates the status of a pull request.

UpdatePullRequestTitle

Replaces the title of a pull request.

UpdateRepositoryDescription

Sets or changes the comment or description for a repository.

UpdateRepositoryName

Renames a repository. The repository name must be unique across the calling AWS