Struct aws_sdk_codecommit::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for CodeCommit_20150413.

This client allows ergonomic access to a CodeCommit_20150413-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn associate_approval_rule_template_with_repository(
&self
) -> AssociateApprovalRuleTemplateWithRepository<C, M, R>

Constructs a fluent builder for the AssociateApprovalRuleTemplateWithRepository operation.

See AssociateApprovalRuleTemplateWithRepository for more information about the operation and its arguments.

pub fn batch_associate_approval_rule_template_with_repositories(
&self
) -> BatchAssociateApprovalRuleTemplateWithRepositories<C, M, R>

Constructs a fluent builder for the BatchAssociateApprovalRuleTemplateWithRepositories operation.

See BatchAssociateApprovalRuleTemplateWithRepositories for more information about the operation and its arguments.

pub fn batch_describe_merge_conflicts(
&self
) -> BatchDescribeMergeConflicts<C, M, R>

Constructs a fluent builder for the BatchDescribeMergeConflicts operation.

See BatchDescribeMergeConflicts for more information about the operation and its arguments.

pub fn batch_disassociate_approval_rule_template_from_repositories(
&self
) -> BatchDisassociateApprovalRuleTemplateFromRepositories<C, M, R>

Constructs a fluent builder for the BatchDisassociateApprovalRuleTemplateFromRepositories operation.

See BatchDisassociateApprovalRuleTemplateFromRepositories for more information about the operation and its arguments.

pub fn batch_get_commits(&self) -> BatchGetCommits<C, M, R>

Constructs a fluent builder for the BatchGetCommits operation.

See BatchGetCommits for more information about the operation and its arguments.

pub fn batch_get_repositories(&self) -> BatchGetRepositories<C, M, R>

Constructs a fluent builder for the BatchGetRepositories operation.

See BatchGetRepositories for more information about the operation and its arguments.

pub fn create_approval_rule_template(
&self
) -> CreateApprovalRuleTemplate<C, M, R>

Constructs a fluent builder for the CreateApprovalRuleTemplate operation.

See CreateApprovalRuleTemplate for more information about the operation and its arguments.

pub fn create_branch(&self) -> CreateBranch<C, M, R>

Constructs a fluent builder for the CreateBranch operation.

See CreateBranch for more information about the operation and its arguments.

pub fn create_commit(&self) -> CreateCommit<C, M, R>

Constructs a fluent builder for the CreateCommit operation.

See CreateCommit for more information about the operation and its arguments.

pub fn create_pull_request(&self) -> CreatePullRequest<C, M, R>

Constructs a fluent builder for the CreatePullRequest operation.

See CreatePullRequest for more information about the operation and its arguments.

pub fn create_pull_request_approval_rule(
&self
) -> CreatePullRequestApprovalRule<C, M, R>

Constructs a fluent builder for the CreatePullRequestApprovalRule operation.

See CreatePullRequestApprovalRule for more information about the operation and its arguments.

pub fn create_repository(&self) -> CreateRepository<C, M, R>

Constructs a fluent builder for the CreateRepository operation.

See CreateRepository for more information about the operation and its arguments.

pub fn create_unreferenced_merge_commit(
&self
) -> CreateUnreferencedMergeCommit<C, M, R>

Constructs a fluent builder for the CreateUnreferencedMergeCommit operation.

See CreateUnreferencedMergeCommit for more information about the operation and its arguments.

pub fn delete_approval_rule_template(
&self
) -> DeleteApprovalRuleTemplate<C, M, R>

Constructs a fluent builder for the DeleteApprovalRuleTemplate operation.

See DeleteApprovalRuleTemplate for more information about the operation and its arguments.

pub fn delete_branch(&self) -> DeleteBranch<C, M, R>

Constructs a fluent builder for the DeleteBranch operation.

See DeleteBranch for more information about the operation and its arguments.

pub fn delete_comment_content(&self) -> DeleteCommentContent<C, M, R>

Constructs a fluent builder for the DeleteCommentContent operation.

See DeleteCommentContent for more information about the operation and its arguments.

pub fn delete_file(&self) -> DeleteFile<C, M, R>

Constructs a fluent builder for the DeleteFile operation.

See DeleteFile for more information about the operation and its arguments.

pub fn delete_pull_request_approval_rule(
&self
) -> DeletePullRequestApprovalRule<C, M, R>

Constructs a fluent builder for the DeletePullRequestApprovalRule operation.

See DeletePullRequestApprovalRule for more information about the operation and its arguments.

pub fn delete_repository(&self) -> DeleteRepository<C, M, R>

Constructs a fluent builder for the DeleteRepository operation.

See DeleteRepository for more information about the operation and its arguments.

pub fn describe_merge_conflicts(&self) -> DescribeMergeConflicts<C, M, R>

Constructs a fluent builder for the DescribeMergeConflicts operation.

See DescribeMergeConflicts for more information about the operation and its arguments.

pub fn describe_pull_request_events(&self) -> DescribePullRequestEvents<C, M, R>

Constructs a fluent builder for the DescribePullRequestEvents operation.

See DescribePullRequestEvents for more information about the operation and its arguments.

pub fn disassociate_approval_rule_template_from_repository(
&self
) -> DisassociateApprovalRuleTemplateFromRepository<C, M, R>

Constructs a fluent builder for the DisassociateApprovalRuleTemplateFromRepository operation.

See DisassociateApprovalRuleTemplateFromRepository for more information about the operation and its arguments.

pub fn evaluate_pull_request_approval_rules(
&self
) -> EvaluatePullRequestApprovalRules<C, M, R>

Constructs a fluent builder for the EvaluatePullRequestApprovalRules operation.

See EvaluatePullRequestApprovalRules for more information about the operation and its arguments.

pub fn get_approval_rule_template(&self) -> GetApprovalRuleTemplate<C, M, R>

Constructs a fluent builder for the GetApprovalRuleTemplate operation.

See GetApprovalRuleTemplate for more information about the operation and its arguments.

pub fn get_blob(&self) -> GetBlob<C, M, R>

Constructs a fluent builder for the GetBlob operation.

See GetBlob for more information about the operation and its arguments.

pub fn get_branch(&self) -> GetBranch<C, M, R>

Constructs a fluent builder for the GetBranch operation.

See GetBranch for more information about the operation and its arguments.

pub fn get_comment(&self) -> GetComment<C, M, R>

Constructs a fluent builder for the GetComment operation.

See GetComment for more information about the operation and its arguments.

pub fn get_comment_reactions(&self) -> GetCommentReactions<C, M, R>

Constructs a fluent builder for the GetCommentReactions operation.

See GetCommentReactions for more information about the operation and its arguments.

pub fn get_comments_for_compared_commit(
&self
) -> GetCommentsForComparedCommit<C, M, R>

Constructs a fluent builder for the GetCommentsForComparedCommit operation.

See GetCommentsForComparedCommit for more information about the operation and its arguments.

pub fn get_comments_for_pull_request(
&self
) -> GetCommentsForPullRequest<C, M, R>

Constructs a fluent builder for the GetCommentsForPullRequest operation.

See GetCommentsForPullRequest for more information about the operation and its arguments.

pub fn get_commit(&self) -> GetCommit<C, M, R>

Constructs a fluent builder for the GetCommit operation.

See GetCommit for more information about the operation and its arguments.

pub fn get_differences(&self) -> GetDifferences<C, M, R>

Constructs a fluent builder for the GetDifferences operation.

See GetDifferences for more information about the operation and its arguments.

pub fn get_file(&self) -> GetFile<C, M, R>

Constructs a fluent builder for the GetFile operation.

See GetFile for more information about the operation and its arguments.

pub fn get_folder(&self) -> GetFolder<C, M, R>

Constructs a fluent builder for the GetFolder operation.

See GetFolder for more information about the operation and its arguments.

pub fn get_merge_commit(&self) -> GetMergeCommit<C, M, R>

Constructs a fluent builder for the GetMergeCommit operation.

See GetMergeCommit for more information about the operation and its arguments.

pub fn get_merge_conflicts(&self) -> GetMergeConflicts<C, M, R>

Constructs a fluent builder for the GetMergeConflicts operation.

See GetMergeConflicts for more information about the operation and its arguments.

pub fn get_merge_options(&self) -> GetMergeOptions<C, M, R>

Constructs a fluent builder for the GetMergeOptions operation.

See GetMergeOptions for more information about the operation and its arguments.

pub fn get_pull_request(&self) -> GetPullRequest<C, M, R>

Constructs a fluent builder for the GetPullRequest operation.

See GetPullRequest for more information about the operation and its arguments.

pub fn get_pull_request_approval_states(
&self
) -> GetPullRequestApprovalStates<C, M, R>

Constructs a fluent builder for the GetPullRequestApprovalStates operation.

See GetPullRequestApprovalStates for more information about the operation and its arguments.

pub fn get_pull_request_override_state(
&self
) -> GetPullRequestOverrideState<C, M, R>

Constructs a fluent builder for the GetPullRequestOverrideState operation.

See GetPullRequestOverrideState for more information about the operation and its arguments.

pub fn get_repository(&self) -> GetRepository<C, M, R>

Constructs a fluent builder for the GetRepository operation.

See GetRepository for more information about the operation and its arguments.

pub fn get_repository_triggers(&self) -> GetRepositoryTriggers<C, M, R>

Constructs a fluent builder for the GetRepositoryTriggers operation.

See GetRepositoryTriggers for more information about the operation and its arguments.

pub fn list_approval_rule_templates(&self) -> ListApprovalRuleTemplates<C, M, R>

Constructs a fluent builder for the ListApprovalRuleTemplates operation.

See ListApprovalRuleTemplates for more information about the operation and its arguments.

pub fn list_associated_approval_rule_templates_for_repository(
&self
) -> ListAssociatedApprovalRuleTemplatesForRepository<C, M, R>

Constructs a fluent builder for the ListAssociatedApprovalRuleTemplatesForRepository operation.

See ListAssociatedApprovalRuleTemplatesForRepository for more information about the operation and its arguments.

pub fn list_branches(&self) -> ListBranches<C, M, R>

Constructs a fluent builder for the ListBranches operation.

See ListBranches for more information about the operation and its arguments.

pub fn list_pull_requests(&self) -> ListPullRequests<C, M, R>

Constructs a fluent builder for the ListPullRequests operation.

See ListPullRequests for more information about the operation and its arguments.

pub fn list_repositories(&self) -> ListRepositories<C, M, R>

Constructs a fluent builder for the ListRepositories operation.

See ListRepositories for more information about the operation and its arguments.

pub fn list_repositories_for_approval_rule_template(
&self
) -> ListRepositoriesForApprovalRuleTemplate<C, M, R>

Constructs a fluent builder for the ListRepositoriesForApprovalRuleTemplate operation.

See ListRepositoriesForApprovalRuleTemplate for more information about the operation and its arguments.

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

Constructs a fluent builder for the ListTagsForResource operation.

See ListTagsForResource for more information about the operation and its arguments.

pub fn merge_branches_by_fast_forward(
&self
) -> MergeBranchesByFastForward<C, M, R>

Constructs a fluent builder for the MergeBranchesByFastForward operation.

See MergeBranchesByFastForward for more information about the operation and its arguments.

pub fn merge_branches_by_squash(&self) -> MergeBranchesBySquash<C, M, R>

Constructs a fluent builder for the MergeBranchesBySquash operation.

See MergeBranchesBySquash for more information about the operation and its arguments.

pub fn merge_branches_by_three_way(&self) -> MergeBranchesByThreeWay<C, M, R>

Constructs a fluent builder for the MergeBranchesByThreeWay operation.

See MergeBranchesByThreeWay for more information about the operation and its arguments.

pub fn merge_pull_request_by_fast_forward(
&self
) -> MergePullRequestByFastForward<C, M, R>

Constructs a fluent builder for the MergePullRequestByFastForward operation.

See MergePullRequestByFastForward for more information about the operation and its arguments.

pub fn merge_pull_request_by_squash(&self) -> MergePullRequestBySquash<C, M, R>

Constructs a fluent builder for the MergePullRequestBySquash operation.

See MergePullRequestBySquash for more information about the operation and its arguments.

pub fn merge_pull_request_by_three_way(
&self
) -> MergePullRequestByThreeWay<C, M, R>

Constructs a fluent builder for the MergePullRequestByThreeWay operation.

See MergePullRequestByThreeWay for more information about the operation and its arguments.

pub fn override_pull_request_approval_rules(
&self
) -> OverridePullRequestApprovalRules<C, M, R>

Constructs a fluent builder for the OverridePullRequestApprovalRules operation.

See OverridePullRequestApprovalRules for more information about the operation and its arguments.

pub fn post_comment_for_compared_commit(
&self
) -> PostCommentForComparedCommit<C, M, R>

Constructs a fluent builder for the PostCommentForComparedCommit operation.

See PostCommentForComparedCommit for more information about the operation and its arguments.

pub fn post_comment_for_pull_request(
&self
) -> PostCommentForPullRequest<C, M, R>

Constructs a fluent builder for the PostCommentForPullRequest operation.

See PostCommentForPullRequest for more information about the operation and its arguments.

pub fn post_comment_reply(&self) -> PostCommentReply<C, M, R>

Constructs a fluent builder for the PostCommentReply operation.

See PostCommentReply for more information about the operation and its arguments.

pub fn put_comment_reaction(&self) -> PutCommentReaction<C, M, R>

Constructs a fluent builder for the PutCommentReaction operation.

See PutCommentReaction for more information about the operation and its arguments.

pub fn put_file(&self) -> PutFile<C, M, R>

Constructs a fluent builder for the PutFile operation.

See PutFile for more information about the operation and its arguments.

pub fn put_repository_triggers(&self) -> PutRepositoryTriggers<C, M, R>

Constructs a fluent builder for the PutRepositoryTriggers operation.

See PutRepositoryTriggers for more information about the operation and its arguments.

pub fn tag_resource(&self) -> TagResource<C, M, R>

Constructs a fluent builder for the TagResource operation.

See TagResource for more information about the operation and its arguments.

pub fn test_repository_triggers(&self) -> TestRepositoryTriggers<C, M, R>

Constructs a fluent builder for the TestRepositoryTriggers operation.

See TestRepositoryTriggers for more information about the operation and its arguments.

pub fn untag_resource(&self) -> UntagResource<C, M, R>

Constructs a fluent builder for the UntagResource operation.

See UntagResource for more information about the operation and its arguments.

pub fn update_approval_rule_template_content(
&self
) -> UpdateApprovalRuleTemplateContent<C, M, R>

Constructs a fluent builder for the UpdateApprovalRuleTemplateContent operation.

See UpdateApprovalRuleTemplateContent for more information about the operation and its arguments.

pub fn update_approval_rule_template_description(
&self
) -> UpdateApprovalRuleTemplateDescription<C, M, R>

Constructs a fluent builder for the UpdateApprovalRuleTemplateDescription operation.

See UpdateApprovalRuleTemplateDescription for more information about the operation and its arguments.

pub fn update_approval_rule_template_name(
&self
) -> UpdateApprovalRuleTemplateName<C, M, R>

Constructs a fluent builder for the UpdateApprovalRuleTemplateName operation.

See UpdateApprovalRuleTemplateName for more information about the operation and its arguments.

pub fn update_comment(&self) -> UpdateComment<C, M, R>

Constructs a fluent builder for the UpdateComment operation.

See UpdateComment for more information about the operation and its arguments.

pub fn update_default_branch(&self) -> UpdateDefaultBranch<C, M, R>

Constructs a fluent builder for the UpdateDefaultBranch operation.

See UpdateDefaultBranch for more information about the operation and its arguments.

pub fn update_pull_request_approval_rule_content(
&self
) -> UpdatePullRequestApprovalRuleContent<C, M, R>

Constructs a fluent builder for the UpdatePullRequestApprovalRuleContent operation.

See UpdatePullRequestApprovalRuleContent for more information about the operation and its arguments.

pub fn update_pull_request_approval_state(
&self
) -> UpdatePullRequestApprovalState<C, M, R>

Constructs a fluent builder for the UpdatePullRequestApprovalState operation.

See UpdatePullRequestApprovalState for more information about the operation and its arguments.

pub fn update_pull_request_description(
&self
) -> UpdatePullRequestDescription<C, M, R>

Constructs a fluent builder for the UpdatePullRequestDescription operation.

See UpdatePullRequestDescription for more information about the operation and its arguments.

pub fn update_pull_request_status(&self) -> UpdatePullRequestStatus<C, M, R>

Constructs a fluent builder for the UpdatePullRequestStatus operation.

See UpdatePullRequestStatus for more information about the operation and its arguments.

pub fn update_pull_request_title(&self) -> UpdatePullRequestTitle<C, M, R>

Constructs a fluent builder for the UpdatePullRequestTitle operation.

See UpdatePullRequestTitle for more information about the operation and its arguments.

pub fn update_repository_description(
&self
) -> UpdateRepositoryDescription<C, M, R>

Constructs a fluent builder for the UpdateRepositoryDescription operation.

See UpdateRepositoryDescription for more information about the operation and its arguments.

pub fn update_repository_name(&self) -> UpdateRepositoryName<C, M, R>

Constructs a fluent builder for the UpdateRepositoryName operation.

See UpdateRepositoryName for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Self

Creates a new client from the service Config.